Update instances.md
Change the second **we** to **be** in the following section.

## Short-Lived Infrastructure

Paragraph

Fortunately, upgrading all your deployments to new BOSH stemcells is a very easy operation and we will definitely **be** returning to this topic soon. For now, you need to know that this process will result in your application processes being stopped (via Monit), the original servers being deleted (via the CPI), and new servers being created to replace them (via the CPI). Any files written arbitrarily to the filesystem will be lost.