Update instances.md (#27)
* Update instances.md

Change the second **we** to **be** in the following section.

## Short-Lived Infrastructure

Paragraph

Fortunately, upgrading all your deployments to new BOSH stemcells is a very easy operation and we will definitely **be** returning to this topic soon. For now, you need to know that this process will result in your application processes being stopped (via Monit), the original servers being deleted (via the CPI), and new servers being created to replace them (via the CPI). Any files written arbitrarily to the filesystem will be lost.

* Update instance.md

Change second **sizing** to **size** in the following section.

## Short-Lived Infrastructure

Another routine operation you will perform that causes instances to be stopped, and the servers deleted and replaced, is resizing or scaling up your instances. BOSH CPIs assume that cloud providers do not know how to **resize** a running server, so they will emulate "resizing" by deleting the original small server and replacing it with a new larger server. Any files written arbitrarily to the filesystem will be lost.